requirements.txt
aiosqlite==0.20.0
annotated-types==0.7.0
anthropic==0.34.2
anyio==4.4.0
boto3==1.35.25
botocore==1.35.25
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fsspec==2024.9.0
gevent==24.2.1
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
qidna==3.8
jiter==0.5.0
jmespath==1.0.1
numpy==2.1.1
packaging==24.1
paho-mqtt==1.6.1
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tokenizers==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
websocket==0.2.1
websockets==13.0.1
y-py==0.6.2
ypy-websocket==0.12.4
zope.event==5.0
zope.interface==7.0.3
